backend/requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.11.7

# Database
chromadb==1.0.15
pymongo==4.13.2

# ML/AI dependencies
llama-cpp-python
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.68
sentence-transformers==5.0.0
rank_bm25==0.2.2
onnxruntime
optimum

# Use CPU-only version of PyTorch to reduce size
--extra-index-url https://download.pytorch.org/whl/cpu

# PaddleOCR 3.x (basic OCR, no need for [all])
paddlepaddle==3.2.0
paddleocr

instructor
openai
python-dateutil